refactor: backport caseworkerOrganizations relocation to notes schema
- Remove caseworkerOrganizations from accounts/schema.py
- Add caseworkerOrganizations to notes/schema.py with user-scoped filtering
- Remove unused CASEWORKER import from accounts/schema.py
- Regenerate schema.graphql
